// src/app/core/models/operaciones/reclamos-model/reclamo.enum.model.ts
export enum EstadoReclamo {
    ABIERTO = 'ABIERTO',
    EN_PROCESO = 'EN_PROCESO',
    RESUELTO = 'RESUELTO',
    CERRADO = 'CERRADO',
    ANULADO = 'ANULADO'
}

export enum TipoReclamacion {
    RECLAMO = 'RECLAMO',
    QUEJA = 'QUEJA'
}

export enum TipoProblema {
    SERVICIO_DEFECTUOSO = 'SERVICIO_DEFECTUOSO',
    CORTE_INCORRECTO = 'CORTE_INCORRECTO',
    BARBERO_INCORRECTO = 'BARBERO_INCORRECTO',
    ATENCION_AL_CLIENTE = 'ATENCION_AL_CLIENTE',
    INCUMPLIMIENTO_HORARIO = 'INCUMPLIMIENTO_HORARIO',
    DEMORA_EN_ATENCION = 'DEMORA_EN_ATENCION',
    RESERVA_INCORRECTA = 'RESERVA_INCORRECTA',
    COBRO_INCORRECTO = 'COBRO_INCORRECTO',
    PRODUCTO_DEFECTUOSO = 'PRODUCTO_DEFECTUOSO',
    PRODUCTO_INCORRECTO = 'PRODUCTO_INCORRECTO',
    PRODUCTO_INCOMPLETO = 'PRODUCTO_INCOMPLETO',
    PRODUCTO_NO_LLEGO = 'PRODUCTO_NO_LLEGO',
    DEMORA_EN_ENTREGA = 'DEMORA_EN_ENTREGA',
    OTRO = 'OTRO'
}

export enum CausaReclamo {
    INCIDENCIA_DEL_PERSONAL = 'INCIDENCIA_DEL_PERSONAL',
    INCIDENCIA_EN_RESERVA = 'INCIDENCIA_EN_RESERVA',
    INCIDENCIA_EN_COBRO = 'INCIDENCIA_EN_COBRO',
    MALA_ATENCION = 'MALA_ATENCION',
    INCUMPLIMIENTO_DE_HORARIO = 'INCUMPLIMIENTO_DE_HORARIO',
    INCIDENCIA_EN_SISTEMA = 'INCIDENCIA_EN_SISTEMA',
    FALTA_DE_STOCK = 'FALTA_DE_STOCK',
    INCIDENCIA_EN_EMPAQUE = 'INCIDENCIA_EN_EMPAQUE',
    INCIDENCIA_EN_DESPACHO = 'INCIDENCIA_EN_DESPACHO',
    FALLA_DEL_PROVEEDOR = 'FALLA_DEL_PROVEEDOR',
    OTRO = 'OTRO'
}

export enum SolucionReclamo {
    REEMBOLSO_TOTAL = 'REEMBOLSO_TOTAL',
    REEMBOLSO_PARCIAL = 'REEMBOLSO_PARCIAL',
    REENVIO_PRODUCTO = 'REENVIO_PRODUCTO',
    CAMBIO_PRODUCTO = 'CAMBIO_PRODUCTO',
    NOTA_CREDITO = 'NOTA_CREDITO',
    DISCULPA_FORMAL = 'DISCULPA_FORMAL',
    BONIFICACION = 'BONIFICACION',
    SIN_SOLUCION = 'SIN_SOLUCION',
    OTRO = 'OTRO'
}

export enum TipoAdjuntoReclamo {
    FOTO = 'FOTO',
    VIDEO = 'VIDEO',
    PDF = 'PDF',
    AUDIO = 'AUDIO',
    CAPTURA = 'CAPTURA',
    OTRO = 'OTRO'
}